telegram_bot/requirements.txt
python-telegram-bot==13.7
psycopg2-binary==2.9.6
redis==4.5.4
python-dotenv==1.0.0
aiogram==2.25.1
asyncpg==0.27.0
aioredis==2.0.1